Main

Documentation

main :: IO ()

main is the starting method for the hangman game

type WordList = [String]

type Wordlist is Syntactic sugar for a list of Strings

allWords :: IO WordList

Reads words from 'dict.txt' into a list dict.txt originates from MacOS's default dictionary

minWordLength :: Int

The minimum length of a word

maxWordLength :: Int

The maximum length of a word

gameWords :: IO WordList

Returns a list of words, based on criteria matching minWordLength, maxWordLength

randomWord :: WordList -> IO String

Randomly chooses a word from a list

randomWord' :: IO String

Chooses a random word from gameWord

data Puzzle

Datatype for Puzzle, contents in order are The word to find, the list of discovered letters, and the list of letters attempted

Constructors

Puzzle String [Maybe Char] [Char]

freshPuzzle :: String -> Puzzle

Generates a new Puzzle datatype, for a new game

charInWord :: Puzzle -> Char -> Bool

Returns True if the character is in the word

alreadyGuessed :: Puzzle -> Char -> Bool

Returns True if the character has already been guessed

renderPuzzleChar :: Maybe Char -> Char

Renders the characters or blank spaces of the discovered word, Nothing is displayed as '_'

fillInCharacter :: Puzzle -> Char -> Puzzle

Places the guessed character into the guess list, and updates the discovered letters if it is in the word

handleGuess :: Puzzle -> Char -> IO Puzzle

Handles the character based on if it was previously chosen or is a character in the word, and prints the appropriate message

gameOver :: Puzzle -> IO ()

Checks if it is game over, based on number of incorrect guesses.

gameWin :: Puzzle -> IO ()

Prints the win message and exits, if all characters are correctly guessed

runGame :: Puzzle -> IO ()

Runs the main portion of the game